Add the ability to feed arrays into SelectDoc 'With' statements
If you feed an array of scalars into a .With(), then you will get a temporary table with a single column labeled "data".

If you feed a struct of data into the .With() then there will be one column per struct field, for every struct field that has a db tag which will be the name of the resulting column. Please do not use nested structs.